Duplicate index test
