+++ package.json
@@ -20,6 +20,7 @@
   "devDependencies": {
     "eslint": "^6.8.0",
     "eslint-plugin-json": "^2.1.1",
+    "glob": "^7.1.6",
     "jest": "^26.0.1",
     "node": "^14.4.0"
   },
